Initial ground work for moving all reverberation/eigenverb code to its own namespace, eigenverb.
